[MIST-632] Create packages for QueryStarter and QueryRemover
