+++ src/main/java/edu/snu/mist/core/task/globalsched/DefaultGlobalSchedGroupInfo.java
@@ -18,6 +18,8 @@ package edu.snu.mist.core.task.globalsched;
 import edu.snu.mist.common.parameters.GroupId;
 import edu.snu.mist.core.parameters.DefaultGroupWeight;
 import edu.snu.mist.core.task.*;
+import edu.snu.mist.core.task.queryRemovers.QueryRemover;
+import edu.snu.mist.core.task.queryStarters.QueryStarter;
 import org.apache.reef.tang.annotations.Parameter;
 
 import javax.inject.Inject;
