+++ src/main/java/edu/snu/mist/core/task/globalsched/GlobalSchedGroupInfo.java
@@ -17,8 +17,8 @@ package edu.snu.mist.core.task.globalsched;
 
 import edu.snu.mist.core.task.ExecutionDags;
 import edu.snu.mist.core.task.OperatorChainManager;
-import edu.snu.mist.core.task.QueryRemover;
-import edu.snu.mist.core.task.QueryStarter;
+import edu.snu.mist.core.task.queryRemovers.QueryRemover;
+import edu.snu.mist.core.task.queryStarters.QueryStarter;
 import org.apache.reef.tang.annotations.DefaultImplementation;
 
 import java.util.List;
