+++ src/main/java/edu/snu/mist/core/task/queryRemovers/MergeAwareQueryRemover.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.core.task;
+package edu.snu.mist.core.task.queryRemovers;
 
 import edu.snu.mist.common.graph.DAG;
 import edu.snu.mist.common.graph.MISTEdge;
+import edu.snu.mist.core.task.*;
 
 import javax.inject.Inject;
 import java.util.Collection;
