+++ src/main/java/edu/snu/mist/core/task/queryStarters/CommonSubDagFinder.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.core.task;
+package edu.snu.mist.core.task.queryStarters;
 
 import edu.snu.mist.common.graph.DAG;
 import edu.snu.mist.common.graph.MISTEdge;
+import edu.snu.mist.core.task.ExecutionVertex;
 import org.apache.reef.tang.annotations.DefaultImplementation;
 
 import java.util.Map;
