+++ src/main/java/edu/snu/mist/core/task/queryStarters/QueryStarterUtils.java
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.core.task;
+package edu.snu.mist.core.task.queryStarters;
 
 import edu.snu.mist.common.graph.DAG;
 import edu.snu.mist.common.graph.GraphUtils;
 import edu.snu.mist.common.graph.MISTEdge;
+import edu.snu.mist.core.task.*;
 
 import java.util.Iterator;
 import java.util.Map;
